Refactor ProductCategory handling and validation

+++ src/main/java/com/froilan/synectix/service/inventory/ProductCategoryManagementService.java
@@ -3,6 +3,7 @@ package com.froilan.synectix.service.inventory;
 import com.froilan.synectix.exception.validation.NotFoundException;
 import com.froilan.synectix.model.Company;
 import com.froilan.synectix.model.dto.request.inventory.ProductCategoryCreateBody;
+import com.froilan.synectix.model.dto.response.product.category.ProductCategoryDTO;
 import com.froilan.synectix.model.inventory.ProductCategory;
 
 import jakarta.persistence.EntityManager;
@@ -30,7 +31,7 @@ public class ProductCategoryManagementService {
     }
 
     @Transactional
-    public void createProductCategory(ProductCategoryCreateBody newProductCategoryBody, String companyUuid)
+    public Integer createProductCategory(ProductCategoryCreateBody newProductCategoryBody, String companyUuid)
             throws EntityNotFoundException, IllegalArgumentException, OptimisticLockingFailureException {
         Company companyReference = entityManager.getReference(Company.class, companyUuid);
         ProductCategory productCategory = ProductCategory.builder()
@@ -39,25 +40,54 @@ public class ProductCategoryManagementService {
                 .description(newProductCategoryBody.getDescription())
                 .company(companyReference)
                 .build();
-        productCategoryRepository.save(productCategory);
+        return productCategoryRepository.save(productCategory).getProductCategoryId();
     }
 
     @Transactional
-    public void deleteProductCategory(Integer id) throws NotFoundException {
+    public boolean deleteProductCategory(Integer id, String companyUuid) throws NotFoundException {
         ProductCategory productCategory = productCategoryRepository.findById(id)
                 .orElseThrow(
                         () -> new EntityNotFoundException("ProductCategory not found with ID: " + id));
+        if (!productCategory.getCompany().getUuid().toString().equals(companyUuid)) {
+            // TODO: Create and throw a custom exception for unauthorized access
+            throw new NotFoundException("ProductCategory not found with ID: " + id);
+        }
         productCategoryRepository.delete(productCategory);
+        return true;
     }
 
-    public List<ProductCategory>  getProductCategories(String companyUuid) {
-        // This method can be implemented to retrieve all product categories associated with a company.
-        // For now, it is left empty as per the original code structure
-        return productCategoryRepository.findAllByCompanyUuid(UUID.fromString(companyUuid));
+    public List<ProductCategoryDTO> getProductCategories(String companyUuid) {
+        return productCategoryRepository.findAllByCompanyUuid(UUID.fromString(companyUuid)).stream()
+                .map(productCategory -> ProductCategoryDTO.builder()
+                        .productCategoryId(productCategory.getProductCategoryId())
+                        .productCategoryCode(productCategory.getProductCategoryCode())
+                        .productCategoryName(productCategory.getProductCategoryName())
+                        .description(productCategory.getDescription())
+                        .build())
+                .toList();
     }
 
-    public ProductCategory getProductCategory(Integer id) throws NotFoundException {
-        return productCategoryRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("ProductCategory not found with ID: " + id));
+    public ProductCategoryDTO getProductCategory(Integer id) throws NotFoundException {
+        ProductCategory productCategory = productCategoryRepository.findById(id).orElseThrow(
+                () -> new NotFoundException("ProductCategory not found with ID: " + id));
+        return ProductCategoryDTO.builder()
+                .productCategoryId(productCategory.getProductCategoryId())
+                .productCategoryCode(productCategory.getProductCategoryCode())
+                .productCategoryName(productCategory.getProductCategoryName())
+                .description(productCategory.getDescription())
+                .build();
+    }
+
+    public Integer updateProductCategory(Integer valueOf, ProductCategoryCreateBody request, String companyUuid) {
+        ProductCategory productCategory = productCategoryRepository.findById(valueOf)
+                .orElseThrow(() -> new NotFoundException("ProductCategory not found with ID: " + valueOf));
+        if (!productCategory.getCompany().getUuid().toString().equals(companyUuid)) {
+            // TODO: Create and throw a custom exception for unauthorized access
+            throw new NotFoundException("ProductCategory not found with ID: " + valueOf);
+        }
+        productCategory.setProductCategoryCode(request.getProductCategoryCode());
+        productCategory.setProductCategoryName(request.getProductCategoryName());
+        productCategory.setDescription(request.getDescription());
+        return productCategoryRepository.save(productCategory).getProductCategoryId();
     }
 }
